Added display font size variable tfiala-fontsize.
Per-machine configurations may now specify tfiala-fontsize in the pre
portion. The font will then be set (via init.el) to DejaVu
Mono-{tfiala-fontsize}.

Modified tfiala-keyboard-use-kinesis usage within init.el to not barf if
that variable is not defined.

Removed an extra copy of the kinesis logic that got merged improperly.