Added mac laptop keyboard right C-modifier support
For a laptop keyboard, the right option is now mapped
to the Control (C-) modifier.

Keyboard-related functions are broken out into the
tfiala-keyboard module and no longer reside in
init.el.

Added the link for snappy.home.el to snappy.local.el.
Recent home network changes modified the network name
of this machine.